Feat/workspace

+++ src/controller/workspace/workspace_controller.ts
@@ -8,12 +8,18 @@ import {
   fromWorkspaceDetailToWorkspaceResponseList,
   UpdateWorkspaceData, WorkspaceControllerI, WorkspaceCreateData, WorkspaceFilter, WorkspaceResponse
 } from "./workspace_interfaces";
+import { RoleRepositoryI } from "@/repository/role_access/role_interfaces";
+import { UserRepositoryI } from "@/repository/user/user_interfaces";
 
 export class WorkspaceController implements WorkspaceControllerI {
   private workspace_repo: WorkspaceRepositoryI
+  private role_access_repo: RoleRepositoryI
+  private user_repo: UserRepositoryI
 
-  constructor(workspace_repo: WorkspaceRepositoryI) {
+  constructor(workspace_repo: WorkspaceRepositoryI, role_access_repo: RoleRepositoryI, user_repo: UserRepositoryI) {
     this.workspace_repo = workspace_repo;
+    this.role_access_repo = role_access_repo;
+    this.user_repo = user_repo;
     this.GetWorkspace = this.GetWorkspace.bind(this);
     this.GetWorkspaceList = this.GetWorkspaceList.bind(this);
     this.DeleteWorkspace = this.DeleteWorkspace.bind(this);
@@ -29,11 +35,30 @@ export class WorkspaceController implements WorkspaceControllerI {
       })
     }
 
-    let checkWorkspace = await this.workspace_repo.getWorkspace(new filterWorkspaceDetail({ name: data.name }));
+    let checkWorkspace = await this.workspace_repo.getWorkspace(new filterWorkspaceDetail({ slug: data.slug }));
     if (checkWorkspace.status_code == StatusCodes.OK) {
       return new ResponseData({
-        message: "you already name already taken by others",
+        message: "the slug is already taken by others",
         status_code: StatusCodes.CONFLICT,
+      })
+    }
+
+    let checkAccount = await this.user_repo.getUser({id: user_id});
+    if (checkAccount.status_code != StatusCodes.OK) {
+      return new ResponseData({
+        message: "user is not found",
+        status_code: StatusCodes.BAD_REQUEST,
+      })
+    }
+
+    let defaultRole = await this.role_access_repo.getRole({
+      default: true, 
+      createDefaultWhenNone: true,
+    });
+    if (!(defaultRole.status_code == StatusCodes.OK || defaultRole.status_code == StatusCodes.CREATED)) {
+      return new ResponseData({
+        message: defaultRole.message,
+        status_code: defaultRole.status_code,
       })
     }
 
@@ -45,7 +70,7 @@ export class WorkspaceController implements WorkspaceControllerI {
       })
     }
 
-    let addMemberResponse = await this.workspace_repo.addMember(createResponse.data!.id!, user_id, "1c351187-97a9-4052-9823-8466849cd3bb");
+    let addMemberResponse = await this.workspace_repo.addMember(createResponse.data!.id!, user_id, defaultRole.data?.id!);
     if (addMemberResponse != StatusCodes.NO_CONTENT) {
       return new ResponseData({
         message: "error to sign user as workspace owner",
@@ -63,6 +88,13 @@ export class WorkspaceController implements WorkspaceControllerI {
   }
 
   async GetWorkspace(filter: WorkspaceFilter): Promise<ResponseData<WorkspaceResponse>> {
+    let errorField = filter.getErrorField()
+    if (errorField) {
+      return new ResponseData({
+        message: errorField,
+        status_code: StatusCodes.BAD_REQUEST,
+      })
+    }
     let checkWorkspace = await this.workspace_repo.getWorkspace(filter.toFilterWorkspaceDetail());
     if (checkWorkspace.status_code == StatusCodes.NOT_FOUND){
       return new ResponseData({
@@ -87,6 +119,13 @@ export class WorkspaceController implements WorkspaceControllerI {
   }
 
   async DeleteWorkspace(filter: WorkspaceFilter): Promise<ResponseData<null>> {
+    let errorField = filter.getErrorField()
+    if (errorField) {
+      return new ResponseData({
+        message: errorField,
+        status_code: StatusCodes.BAD_REQUEST,
+      })
+    }
     const deleteResponse = await this.workspace_repo.deleteWorkspace(filter);
     if (deleteResponse == StatusCodes.NOT_FOUND) {
       return new ResponseData({
@@ -101,6 +140,7 @@ export class WorkspaceController implements WorkspaceControllerI {
   }
 
   async UpdateWorkspace(filter: WorkspaceFilter, data: UpdateWorkspaceData): Promise<ResponseData<null>> {
+    let errorField = filter.getErrorField()
     if (filter.isEmpty()) {
       return new ResponseData({
         message: "you need filter to update",
@@ -113,8 +153,25 @@ export class WorkspaceController implements WorkspaceControllerI {
         status_code: StatusCodes.NOT_FOUND,
       })
     }
+    if (errorField) {
+      return new ResponseData({
+        message: errorField,
+        status_code: StatusCodes.BAD_REQUEST,
+      })
+    }
 
-    if (filter.id || filter.user_id_owner) {
+    if (filter.user_id_owner) {
+      let workspaceByUser = await this.GetWorkspace(new WorkspaceFilter({user_id_owner: filter.user_id_owner}));
+      if (workspaceByUser.status_code != StatusCodes.OK) {
+        return new ResponseData({
+          message: workspaceByUser.message,
+          status_code: StatusCodes.NOT_FOUND,
+        })
+      }
+      filter.id = workspaceByUser.data?.id!
+    }
+
+    if (filter.id) {
       let currentWorkspace = await this.workspace_repo.getWorkspace(new filterWorkspaceDetail({ id: filter.id }));
       if (currentWorkspace.status_code == StatusCodes.NOT_FOUND) {
         return new ResponseData({
